requirements.txt

django
djangorestframework
djangorestframework-simplejwt
drf-spectacular